Raw logging: dDetermine features per sample
Define per sample whether or not process accounting is active and
whether or not disk I/O statitics are maintained.
In previous versions, this was determined for the entire log file
which implied that it was impossible to append to an existing logfile
when one of these features had changed.