Solved segmentation fault in case of threadview
Sometimes a process is considered to be inactive while one of its
threads appears to be active. This situation could lead to a
segmentation fault in case of threadview of only active tasks.
It has been solved by marking the inactive process as active when
later on a thread is discovered that appears to be active.