Introduce Docker container statistics
Maintain the Docker container id (12 positions) per
process and per thread. This container id will be shown
in various lists as CID, but only if at least one Docker
container is present.
Furthermore, it is possible to select all processes
and threads that are related to a specific container id
(keystroke 'J') and to accumulate the resource consumption
for all processes/threads per container (keystroke 'j').