Fix NFS parseable output documentation
Order of the elements in the documentation for the NFS rows in parseable output mismatches the order of the parameters of the printf(...) function in the function "void
print_NFS(char *hp, struct sstat *ss, struct tstat *ps, int nact)" in parseable.c file.

Note : I did not check what the exact meaning of the ss->nfs.server.netcnt and ss->nfs.server.nettcpcon. So, don't forget to modify the row n°1125 and n°1128 in my commit.