atopacctd: provide strerror for "receive NETLINK family"
I tried running atopacctd in my systemd-nspawn container.  Let's provide a
textual error description, which is easier to look up (google).

Before:
    receive NETLINK family, errno -2
After:
    get NETLINK family for taskstats: No such file or directory

Also clarifies where the failure is: we do have NETLINK, but we don't have
the NETLINK family for taskstats.

Google won't find this commit message.  But for reference, the kernel code
which generates this error in ctrl_getfamily() is:

	if (!res->netnsok && !net_eq(genl_info_net(info), &init_net)) {
		/* family doesn't exist here */
		return -ENOENT;
	}

Amusingly, it is indeed permitted to send
TASKSTATS_CMD_ATTR_REGISTER_CPUMASK inside a container which shares the
init netns (which of course does not share the pid namespace).  Dunno if
this would considered an info leak, but at least acct() fails in this
environment.