atopacctd: fix service readiness, avoid hang
The parent process waited to be killed.  The problem is that unexpected
termination of the child will cause the parent to hang.

This fixes a hang when running atopacctd in a systemd-nspawn container.
It now exits immediately with an error.  The problem was that
netlink_open() calls exit() on failure, without killing the parent process.

We also now avoid any possibility of killing an unexpected process, in case
the *parent* process is terminated unexpectedly and its PID is re-used.
These are two examples of why message-passing with a pipe is preferable to
working with kill().

A systemd-specific comment is modified.  AFAICT, a service readiness
protocol is always necessary when you have another service depending on you
(and you do not use socket activation).  You might think atopacctd would be
a candidate for such...  In which case, systemd checking for a pidfile ASAP
just provides a sanity check that was missing from historical
implementations, highlighting that many daemons were not strictly
implementing a service readiness protocol, and hence many boot "sequences"
were actually a pile of race conditions.

Official systemd propaganda for this is currently available[1]; it
explicitly suggests the pipe approach.
[1] https://www.freedesktop.org/software/systemd/man/daemon.html

I'm not really clear from my reading, as to what extent this applied to
atopacctd.  For all I know, atop doesn't mind if atopacctd is started
after it and atop manages to connect immediately when atopacctd is
ready.  However IMO this commit is a well-defined approach, and it solves
a genuine annoyance.