atopacctd: light hack around missing cleanup in error path
Imaging installing atop in a systemd-nspawn container,  enabling atopacctd,
and rebooting.  atopacctd will fail to start in the container.  However,
if you try to run it manually, in order to see the error message or to
troubleshoot it with `strace`, it will then start failing with a
*different* error.

    /var/run/pacct_shadow.d: File exists

This is an unfortunate obstacle to troubleshooting.  Let's fix it...

Currently we don't have code to clean `pacct_shadow.d/` in the event of
atopacctd crashing.  And the failure during startup behaves similar to
a crash.

This commit provides a workaround.  If `pacct_shadow.d` already exists,
then ignore the mkdir() failure.  If it's not a directory like we need,
we'll notice a failure later on anyway.