atopacct.service: don't run on containers
Let's skip on containers here, at least for now.
Currently, Linux Containers don't support process accounting
(and failed services are painful for Debian-style packages which start
the service at install time).

Quietly ignoring a service like this, is a big part of how systemd manages
to run standard distributions inside a container (e.g. skipping
systemd-udev).  (The earlier approach with LXC was for the container to
configure^Wpatch the distribution into submission... it makes me not
inclined to feel very guilty for not having an idea for the _sysvinit_
script).

I'm not sure how useful atop is in containers in practice... however it can
be used as a cheaper way to inspect atop scripts etc. used by a specific
distribution, or developing configuration management to install atop.
I believe there's a non-zero population who have used systemd-nspawn as
"chroot on steroids."