Compilation warnings - String Manipulation
Changed the size of path variable using system constant
Implemented data size check during message creation to avoid buffer overflow and problems when displaying messages.

Now negative numbers are presented as 0(zero) and to big numbers are presented as maximum value(ex.9999). This will avoid problems in case of invalid or unexpected data.

This was tested compiling in 32 and 64 bits.