Try to set the highest OOM priority for atop
We usually expect that atop has highest OOM priority. when the system
runs out of memory, atop still could record necessary infomation.
Then we are able to analyze the OOM situation afterwards.

The orignal verion of this feature: read config from atoprc, then
apply to atop process. But Gerlof considers that just trying to set
the highest priority is better. So rewrite this as Gerlof's advise.

Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>