pricumproc: fix using but without initialization
The "permissables[j]->doconvert()" was added for make_sys_prints()
in former commit 09935eb to check the valid counter, but it also
introduces a using without initialization issue.

In current code, 'as->percputot' is actually assigned in prisyst(),
but this value is used earlier in sysprt_CPUSYS() when this function
is called by pricumproc() => make_sys_prints(). In this way, a
'buffer overflow detected' error message will be triggered via
___sprintf_chk() while compiling.

Fix this by assigning for 'as->percputot' earlier in pricumproc().

Signed-off-by: Fei Li <lifei.shirley@bytedance.com>