mangle CMD field by external helper
New pidcmd option:
The  external  program to translate PID to CMD.
If this keyword is supplied, CMD line is obtained
by running this program. The program must read
newline- terminated  PID from stdin  and write
short newline-terminated string to be used instead
of CMD to stdout, in line-buffered mode.
This may be used to quickly distinguish between
processes of the same executable.
See pidtocmd.pl for example.