Twin mode: live measurement with review option
In twin mode atop spawns into two processes:
- lower level process that gathers the counters and writes
  them to a temporary raw file, and
- upper level process that reads the counters from the raw file
  and presents them to the user.

The reading of the upper level process keeps in pace with the
written samples of the lower level process for live measurements.
However, when pressing the 'r' (reset to measurement begin),
the 'b' (branch to time stamp), or the 'T' (previous sample),
the upper level process implicitly pauses with the possibility
to view previous samples. The 'z' (explicit pause) can also be
used to pause the live measurement.
When pressing the 'z' (continue after pause) the live measurement
will be continued.