Introduce -I flag (solves issue #309 and #68)
With the -I flag the translation from a UID/GID to a name will be
suppressed, so all UIDs and GIDs will be shown as numbers.
This can be useful to view a raw log file from another system
(to avoid wrong translations using info from the local system).

Besides, when a user name or group name is longer than 8 characters
it will automatically be presented as a number instead of showing
a truncated name (even if the -I flag has not been used).