Fix security vulnerability CVE-2025-31160 (#334)
Atop will not connect to the TCP port of 'atopgpud' daemon any more
by default.  The flag -k can be used explicitly when 'atopgpud' is
active. Also the code to parse the received strings is improved to
avoid future issues with heap corruption.

The flag -K has been implemented to connect to netatop/netatop-bpf.