Allow samples with same timestamp in rawlog (#346)
When atop is terminated and restarted right away (like in the
case of 'systemctl restart atop') two samples might get same timestamp.
This should be allowed when extending an existing raw log.

Apart from that, only a check is done on the cgroup version (v2)
of the existing raw log instead of checking on all features.