Allow sorting of process list on any column (#352)
Allow the process list to be sorted on any column, instead of only
cpu (C), memory (M), disk (D), network (N) or gpu (E) utilization
percentage. With the key 'O' a column name can be entered (case
insensitive) that is part of the current process view -- either
visible or invisible -- and a choice can be made to sort ascending
or descending. The current sort column (if visible) will be
highlighted.

The key 'A' (automatically sort on the most utilized resource)
is not supported any more.