Correct sorting of numerical UID/GID (#352)
When the -I flag is used or when a numerical UID/GID can not be
converted into a user/group name, the numerical UID/GID should
be sorted correctly. This means that e.g. UID 31 should be sorted
before UID 181 in case of ascending order and not behind (because
it starts with '3' compared to '1').